refactor(app): Group all interfaces in a package
