+++ app/interfaces/connector.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type Connector interface {
+	Boot()
+}
